Added the random additions of wide moves.
This was the crux of the issue raised in <https://github.com/1Codealot/CLI_Timer/issues/27>.

The random number range for the wide moves is n*3 to n*4, where n is the number layers in the puzzle.

This will be the same for 5 to 7x7 and 3 wide moves will be counted the same as regular wide moves.